+++ src/IkaAgent.cpp
@@ -122,7 +122,8 @@ int IkaAgent::step(double time, double step_size, osi3::SensorView &sensor_view,
                    osi3::TrafficCommand &traffic_command,
                    osi3::TrafficUpdate &traffic_update,
                    setlevel4to5::DynamicsRequest &dynamic_request) {
-  int id = sensor_view.host_vehicle_id().value();
+  uint64_t id = sensor_view.host_vehicle_id().value();
+
 
   std::cout << "---------- time: " << time << " ---------- id: " << id << " ----------" << std::endl;
